Tema13 - Despliegues y funcionalidades avanzadas/Clase38_14-01-2026/requirements.txt
django
djangorestframework
django-cors-headers
djangorestframework-simplejwt
python-dotenv
requests